requirements.txt
# Polymarket Insider Sentinel - Dependencies
web3[async]>=7.0.0
polars>=1.0.0
aiosqlite>=0.20.0
python-dotenv>=1.0.0
aiohttp>=3.9.0
eth-abi>=5.0.0
